refactor(web): fetch cached quick access channels with tanstack
