+++ apps/web/src/lib/queries/channel/channels.ts
@@ -1,6 +1,7 @@
 import { analytics } from '@app/lib/analytics';
 import { toast } from '@core/component/Toast/Toast';
 import { throwOnErr } from '@core/util/result';
+import type { CacheHost } from '@graphql-cache/index';
 import { queryClient } from '@queries/client';
 import { type MutationCallbacks, withCallbacks } from '@queries/utils';
 import type { ApiChannelWithLatest } from '@service-storage/channel-list-types';
@@ -14,6 +15,10 @@ import type { CreateChannelResponse } from '@service-storage/generated/schemas/c
 import type { PatchChannelRequest } from '@service-storage/generated/schemas/patchChannelRequest';
 import { useMutation, useQuery } from '@tanstack/solid-query';
 import { invalidateChannelParticipants } from './channel-participants';
+import {
+  type CachedGraphqlChannel,
+  readCachedGraphqlChannels,
+} from './graphql';
 import { channelKeys } from './keys';
 
 const CHANNEL_LIST_PAGE_SIZE = 100;
@@ -43,6 +48,20 @@ export function useListChannelsQuery() {
   return useQuery(() => ({
     queryKey: channelKeys.listChannels.queryKey,
     queryFn: ({ signal }) => fetchAllChannels(signal),
+  }));
+}
+
+/** Reads the recent Quick Access channel list from the normalized GraphQL cache. */
+export function useCachedGraphqlChannelsQuery(cacheHost?: CacheHost) {
+  return useQuery(() => ({
+    queryKey: channelKeys.quickAccessGraphql(cacheHost?.clientId ?? 'disabled')
+      .queryKey,
+    queryFn: async (): Promise<CachedGraphqlChannel[]> => {
+      if (!cacheHost) return [];
+      return await readCachedGraphqlChannels(cacheHost);
+    },
+    enabled: cacheHost !== undefined,
+    staleTime: Infinity,
   }));
 }
