+++ apps/web/src/lib/queries/history/graphql.ts
@@ -6,8 +6,6 @@ import {
   selectRecords,
 } from '@graphql-cache/index';
 import {
-  type GraphqlChannelQuickAccessFieldsFragment,
-  GraphqlChannelQuickAccessFieldsFragmentDoc,
   type GraphqlChatQuickAccessNameFragment,
   GraphqlChatQuickAccessNameFragmentDoc,
   type GraphqlDocumentQuickAccessNameFragment,
@@ -29,34 +27,6 @@ const HISTORY_TYPENAMES = [
 ] as const;
 
 type HistoryTypename = (typeof HISTORY_TYPENAMES)[number];
-
-/** Minimal cached GraphQL channel fields used by Quick Access. */
-export type CachedGraphqlChannel = {
-  id: string;
-  name: string;
-  ownerId: string;
-  channelType: 'direct_message' | 'private' | 'public' | 'team';
-  participantIds: string[];
-  createdAt: string;
-  updatedAt: string;
-  viewedAt?: string;
-  interactedAt?: string;
-};
-
-function normalizeChannelType(
-  channelType: string
-): CachedGraphqlChannel['channelType'] {
-  switch (channelType.toLowerCase()) {
-    case 'direct_message':
-      return 'direct_message';
-    case 'private':
-      return 'private';
-    case 'team':
-      return 'team';
-    default:
-      return 'public';
-  }
-}
 
 function nameSelection(
   typename: HistoryTypename
@@ -163,58 +133,6 @@ export async function materializeCachedGraphqlHistoryItems(
       : undefined;
     return item ? [item] : [];
   });
-}
-
-/** Materializes supplied cached GraphQL channel search hits. */
-export async function materializeCachedGraphqlChannels(
-  cacheHost: Pick<CacheHost, 'readRecordsByKeys'>,
-  documents: SearchDocumentWire[]
-): Promise<CachedGraphqlChannel[]> {
-  const channelDocuments = documents.filter((document) =>
-    document.recordKey.startsWith('GraphqlSoupChannel:')
-  );
-  if (channelDocuments.length === 0) return [];
-
-  const records = await readRecordsByKeys(
-    cacheHost,
-    selectRecords(GraphqlChannelQuickAccessFieldsFragmentDoc),
-    channelDocuments.map((document) => document.recordKey)
-  );
-
-  return records.flatMap(({ recordKey, record }) => {
-    const channel = record as GraphqlChannelQuickAccessFieldsFragment;
-    if (channel.__typename !== 'GraphqlSoupChannel') return [];
-    const separator = recordKey.indexOf(':');
-    if (separator < 0) return [];
-
-    return [
-      {
-        id: recordKey.slice(separator + 1),
-        name: channel.name ?? '',
-        ownerId: channel.ownerId,
-        channelType: normalizeChannelType(channel.channelType),
-        participantIds: channel.participants.map(({ userId }) => userId),
-        createdAt: channel.createdAt,
-        updatedAt: channel.updatedAt,
-        viewedAt: channel.viewedAt ?? undefined,
-        interactedAt: channel.interactedAt ?? undefined,
-      },
-    ];
-  });
-}
-
-/** Reads the bounded recent channel list used for empty Quick Access menus. */
-export async function readCachedGraphqlChannels(
-  cacheHost: Pick<CacheHost, 'search' | 'readRecordsByKeys'>,
-  limit = 50
-): Promise<CachedGraphqlChannel[]> {
-  const page = await cacheHost.search({
-    profile: 'quick-access-v1',
-    buckets: ['channel', 'dm'],
-    query: '',
-    limit,
-  });
-  return materializeCachedGraphqlChannels(cacheHost, page.documents);
 }
 
 /** Reads a bounded recent history through the indexed search projection, then
